+++ bonus/src/set_image.c
@@ -6,7 +6,7 @@
 /*   By: asuc <asuc@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/30 15:24:42 by bgoron            #+#    #+#             */
-/*   Updated: 2024/07/02 04:46:11 by asuc             ###   ########.fr       */
+/*   Updated: 2024/07/06 16:13:36 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,12 +36,12 @@ void	print_first_background(t_data *data)
 	ceiling_color = argb_to_int(g->wall_sprite.ceiling_color);
 	floor_color = argb_to_int(g->wall_sprite.floor_color);
 	x = 0;
-	while (WIN_WIDTH > x)
+	while ((int)WIN_WIDTH > x)
 	{
 		y = 0;
-		while (WIN_HEIGHT > y)
+		while ((int)WIN_HEIGHT > y)
 		{
-			if (y < WIN_HEIGHT / 2)
+			if (y < (int)WIN_HEIGHT / 2)
 				mlx_set_image_pixel \
 				(g->mlx, g->game.img_background, x, y, ceiling_color);
 			else
